requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
openai-whisper
torch
torchaudio
edge-tts
piper-tts
numpy
opencv-python
moviepy
pydub
soundfile
python-dotenv
pydantic
librosa
httpx
ollama
